Revert "FEATURE: Desktop notification under Tauri app"
Not needed anymore.

This reverts commit 7f100c2d2bb7e77c8c57b6b9fdddbc75b04797d9.
